First step of refactoring of the TerminaOutputDevice
